-enh seperated naitive test into seperete files
