+++ SensorCalibration/test/test_desktop/testMain.cpp
@@ -1,0 +1,18 @@
+#define DOCTEST_CONFIG_IMPLEMENT // REQUIRED: Enable custom main()
+#include <doctest.h>
+
+int main(int argc, char **argv)
+{
+    doctest::Context context;
+
+    // BEGIN:: PLATFORMIO REQUIRED OPTIONS
+    context.setOption("success", true); // Report successful tests
+    context.setOption("no-exitcode",
+                      true); // Do not return non-zero code on failed test case
+    // END:: PLATFORMIO REQUIRED OPTIONS
+
+    // YOUR CUSTOM DOCTEST OPTIONS
+
+    context.applyCommandLine(argc, argv);
+    return context.run();
+}
